requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.1
data==0.4
decorator==5.1.1
funcsigs==1.0.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
selenium==4.25.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0